refactor player.h and player.cpp
move some code in new files to simplify player code a bit.
