+++ server/libFunq/libFunq.pro
@@ -12,6 +12,10 @@ INSTALLS += target
 include(../protocole/protocole.pri)
 
 SOURCES += player.cpp \
+    player_utils.cpp \
+    player_commands_itemmodel.cpp \
+    player_commands_gitems.cpp \
+    player_commands_quickitems.cpp \
     funq.cpp \
     objectpath.cpp \
     dragndropresponse.cpp \
@@ -19,6 +23,7 @@ SOURCES += player.cpp \
     pick.cpp
 
 HEADERS += player.h \
+    player_utils.h \
     funq.h \
     objectpath.h \
     dragndropresponse.h \
