+++ server/libFunq/shortcutresponse.cpp
@@ -34,6 +34,7 @@ knowledge of the CeCILL v2.1 license and that you accept its terms.
 
 #include "shortcutresponse.h"
 #include "player.h"
+#include "player_utils.h"
 #include <QApplication>
 #include <QKeyEvent>
